fix : novelRoom id 필드 null 오류 수정
# Conflicts:
#	src/main/java/com/example/ku_novel/client/ui/HomeUI.java

+++ src/main/java/com/example/ku_novel/common/Message.java
@@ -28,7 +28,7 @@ public class Message {
     private int votingDuration;
     private int submissionDuration;
     private List<Message> novelRooms;
-    private Long novelRoomId;
+    private Integer novelRoomId;
     private String novelRoomTitle;
     private String novelRoomDescription;
     private String novelRoomStatus;
